Add SRW LTSA Package Email
* Split Email Service into StatusEmailService and EmailService
* Update Email Service to allow sending of attachments
