+++ services/apps/alcs/src/portal/notification-submission/notification-submission.controller.ts
@@ -10,8 +10,10 @@ import {
   Req,
   UseGuards,
 } from '@nestjs/common';
+import { PARENT_TYPE } from '../../alcs/card/card-subtask/card-subtask.dto';
 import { NOTIFICATION_STATUS } from '../../alcs/notification/notification-submission-status/notification-status.dto';
 import { PortalAuthGuard } from '../../common/authorization/portal-auth-guard.service';
+import { EmailService } from '../../providers/email/email.service';
 import { User } from '../../user/user.entity';
 import { GenerateSrwDocumentService } from '../pdf-generation/generate-srw-document.service';
 import { NotificationSubmissionValidatorService } from './notification-submission-validator.service';
@@ -28,6 +30,7 @@ export class NotificationSubmissionController {
     private notificationSubmissionService: NotificationSubmissionService,
     private notificationValidationService: NotificationSubmissionValidatorService,
     private generateSrwDocumentService: GenerateSrwDocumentService,
+    private emailService: EmailService,
   ) {}
 
   @Get()
@@ -167,14 +170,33 @@ export class NotificationSubmissionController {
   }
 
   private async generatePdf(submission: NotificationSubmission, user: User) {
-    await this.generateSrwDocumentService.generateAndAttach(
-      submission.fileNumber,
-      user,
-    );
+    const savedDocument =
+      await this.generateSrwDocumentService.generateAndAttach(
+        submission.fileNumber,
+        user,
+      );
 
-    await this.notificationSubmissionService.updateStatus(
-      submission.uuid,
-      NOTIFICATION_STATUS.ALC_RESPONSE_SENT,
-    );
+    if (savedDocument) {
+      const templateData =
+        await this.notificationSubmissionService.generateSrwEmailData(
+          submission,
+          savedDocument,
+        );
+
+      await this.emailService.sendEmail({
+        to: [templateData.to],
+        body: templateData.html,
+        subject: `Agricultural Land Commission SRW${submission.fileNumber} (${submission.applicant})`,
+        parentType: PARENT_TYPE.NOTIFICATION,
+        parentId: templateData.parentId,
+        cc: templateData.cc,
+        attachments: [savedDocument.document],
+      });
+
+      await this.notificationSubmissionService.updateStatus(
+        submission.uuid,
+        NOTIFICATION_STATUS.ALC_RESPONSE_SENT,
+      );
+    }
   }
 }
