+++ services/templates/emails/decision-released/application.template.ts
@@ -1,7 +1,7 @@
 import { MJMLParseResults } from 'mjml-core';
+import { StatusUpdateEmail } from '../../../apps/alcs/src/providers/email/status-email.service';
 import { EmailTemplateService } from '../../../libs/common/src/email-template-service/email-template.service';
 import { header, footer, notificationOnly, portalButton } from '../partials';
-import { StatusUpdateEmail } from '../../../apps/alcs/src/providers/email/email.service';
 
 type DecisionReleasedStatusEmail = StatusUpdateEmail & {
   decisionDate: number;
